Remove workflow barrel and rename shared to _shared
Committed-By-Agent: codex
