--- apps/service/src/temporal/workflows.ts
@@ -1,11 +1,0 @@
-export {
-  configQuery,
-  deleteSignal,
-  stateQuery,
-  statusQuery,
-  stripeEventSignal,
-  updateSignal,
-} from './workflows/shared.js'
-export type { WorkflowStatus } from './workflows/shared.js'
-export { pipelineWorkflow } from './workflows/pipeline.js'
-export { pipelineGoogleSheetsWorkflow } from './workflows/pipeline-google-sheets.js'
