+++ apps/service/src/temporal/workflows/pipeline-google-sheets.ts
@@ -17,7 +17,7 @@ import {
   updateSignal,
   WorkflowStatus,
   writeGoogleSheetsFromQueue,
-} from './shared.js'
+} from './_shared.js'
 import { CONTINUE_AS_NEW_THRESHOLD, deepEqual, EVENT_BATCH_SIZE } from '../../lib/utils.js'
 
 export interface PipelineGoogleSheetsWorkflowOpts {
